TO-RESOLVE
This reflects the merge conflict resolutions with the previous merge
commit as rebased from Git for Windows' `master` on top of upstream.
Once the conflicting patch series make it to an official Git version,
Git for Windows' `master` will need to take these merge conflict
resolutions in a more permanent way.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>